Separate add, sub, sum from test_onnx
