+++ crates/burn-import/onnx-tests/tests/node_tests.rs
@@ -1,0 +1,13 @@
+/// Include specified models in the `model` directory in the target directory.
+#[macro_export]
+macro_rules! include_models {
+    ($($model:ident),*) => {
+        $(
+            // Allow type complexity for generated code
+            #[allow(clippy::type_complexity)]
+            pub mod $model {
+                include!(concat!(env!("OUT_DIR"), concat!("/model/", stringify!($model), ".rs")));
+            }
+        )*
+    };
+}
