+++ crates/burn-import/onnx-tests/tests/sub/mod.rs
@@ -1,0 +1,41 @@
+// Include the models for this node type
+use crate::include_models;
+include_models!(sub, sub_int);
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use burn::tensor::{Int, Tensor, TensorData};
+
+    type Backend = burn_ndarray::NdArray<f32>;
+
+    #[test]
+    fn sub_scalar_from_tensor_and_tensor_from_tensor() {
+        // Initialize the model with weights (loaded from the exported file)
+        let model: sub::Model<Backend> = sub::Model::default();
+
+        let device = Default::default();
+        // Run the model
+        let input = Tensor::<Backend, 4>::from_floats([[[[1., 2., 3., 4.]]]], &device);
+        let scalar = 3.0f64;
+        let output = model.forward(input, scalar);
+        let expected = TensorData::from([[[[-12f32, -13., -14., -15.]]]]);
+
+        output.to_data().assert_eq(&expected, true);
+    }
+
+    #[test]
+    fn sub_scalar_from_int_tensor_and_int_tensor_from_tensor() {
+        // Initialize the model with weights (loaded from the exported file)
+        let model: sub_int::Model<Backend> = sub_int::Model::default();
+
+        let device = Default::default();
+        // Run the model
+        let input = Tensor::<Backend, 4, Int>::from_ints([[[[1, 2, 3, 4]]]], &device);
+        let scalar = 3;
+        let output = model.forward(input, scalar);
+        let expected = TensorData::from([[[[-12i64, -12, -12, -12]]]]);
+
+        output.to_data().assert_eq(&expected, true);
+    }
+}
