+++ crates/burn-import/onnx-tests/tests/test_mod.rs
@@ -1,0 +1,11 @@
+#![no_std]
+
+extern crate alloc;
+
+// Make macro definitions visible
+mod node_tests;
+
+// Import individual node modules
+pub mod add;
+pub mod sub;
+pub mod sum;
